Refactor waypoint migrations
