+++ c2corg_api/scripts/migration/documents/waypoints/__init__.py
@@ -1,0 +1,1 @@
+# package
