+++ c2corg_api/scripts/migration/documents/waypoints/products.py
@@ -1,26 +1,11 @@
-from c2corg_api.models.document import DocumentGeometry, \
-    ArchiveDocumentGeometry
-from c2corg_api.models.waypoint import Waypoint, ArchiveWaypoint, \
-    WaypointLocale, ArchiveWaypointLocale
-from c2corg_api.scripts.migration.documents.document import MigrateDocuments
+from c2corg_api.scripts.migration.documents.waypoints.waypoint import \
+    MigrateWaypoints
 
 
-class MigrateProducts(MigrateDocuments):
+class MigrateProducts(MigrateWaypoints):
 
     def get_name(self):
         return 'products'
-
-    def get_model_document(self, locales):
-        return WaypointLocale if locales else Waypoint
-
-    def get_model_archive_document(self, locales):
-        return ArchiveWaypointLocale if locales else ArchiveWaypoint
-
-    def get_model_geometry(self):
-        return DocumentGeometry
-
-    def get_model_archive_geometry(self):
-        return ArchiveDocumentGeometry
 
     def get_count_query(self):
         return (
@@ -64,24 +49,6 @@ class MigrateProducts(MigrateDocuments):
             url=document_in.url
         )
 
-    def get_document_archive(self, document_in, version):
-        doc = self.get_document(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
-    def get_document_geometry(self, document_in, version):
-        return dict(
-            document_id=document_in.id,
-            id=document_in.id,
-            version=version,
-            geom=document_in.geom
-        )
-
-    def get_document_geometry_archive(self, document_in, version):
-        doc = self.get_document_geometry(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
     def get_document_locale(self, document_in, version):
         # TODO extract summary
         return dict(
@@ -94,9 +61,6 @@ class MigrateProducts(MigrateDocuments):
             access=document_in.access,
             access_period=document_in.hours
         )
-
-    def get_document_locale_archive(self, document_in, version):
-        return self.get_document_locale(document_in, version)
 
     product_types = {
         '1': 'farm_sale',
