+++ c2corg_api/scripts/migration/documents/waypoints/sites.py
@@ -1,27 +1,12 @@
 # coding=utf-8
-from c2corg_api.models.document import DocumentGeometry, \
-    ArchiveDocumentGeometry
-from c2corg_api.models.waypoint import Waypoint, ArchiveWaypoint, \
-    WaypointLocale, ArchiveWaypointLocale
-from c2corg_api.scripts.migration.documents.document import MigrateDocuments
+from c2corg_api.scripts.migration.documents.waypoints.waypoint import \
+    MigrateWaypoints
 
 
-class MigrateSites(MigrateDocuments):
+class MigrateSites(MigrateWaypoints):
 
     def get_name(self):
         return 'sites'
-
-    def get_model_document(self, locales):
-        return WaypointLocale if locales else Waypoint
-
-    def get_model_archive_document(self, locales):
-        return ArchiveWaypointLocale if locales else ArchiveWaypoint
-
-    def get_model_geometry(self):
-        return DocumentGeometry
-
-    def get_model_archive_geometry(self):
-        return ArchiveDocumentGeometry
 
     def get_count_query(self):
         return (
@@ -95,24 +80,6 @@ class MigrateSites(MigrateDocuments):
                 document_in.best_periods, MigrateSites.best_periods)
         )
 
-    def get_document_archive(self, document_in, version):
-        doc = self.get_document(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
-    def get_document_geometry(self, document_in, version):
-        return dict(
-            document_id=document_in.id,
-            id=document_in.id,
-            version=version,
-            geom=document_in.geom
-        )
-
-    def get_document_geometry_archive(self, document_in, version):
-        doc = self.get_document_geometry(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
     def get_document_locale(self, document_in, version):
         # TODO extract summary
         return dict(
@@ -124,9 +91,6 @@ class MigrateSites(MigrateDocuments):
             description=self.get_description(document_in),
             access=document_in.pedestrian_access,
         )
-
-    def get_document_locale_archive(self, document_in, version):
-        return self.get_document_locale(document_in, version)
 
     def get_climbing_types(self, document_in):
         """ v6 did not differentiate between indoor and outdoor types, so this
