+++ c2corg_api/scripts/migration/documents/waypoints/summit.py
@@ -1,26 +1,11 @@
-from c2corg_api.models.document import DocumentGeometry, \
-    ArchiveDocumentGeometry
-from c2corg_api.models.waypoint import Waypoint, ArchiveWaypoint, \
-    WaypointLocale, ArchiveWaypointLocale
-from c2corg_api.scripts.migration.documents.document import MigrateDocuments
+from c2corg_api.scripts.migration.documents.waypoints.waypoint import \
+    MigrateWaypoints
 
 
-class MigrateSummits(MigrateDocuments):
+class MigrateSummits(MigrateWaypoints):
 
     def get_name(self):
         return 'summits'
-
-    def get_model_document(self, locales):
-        return WaypointLocale if locales else Waypoint
-
-    def get_model_archive_document(self, locales):
-        return ArchiveWaypointLocale if locales else ArchiveWaypoint
-
-    def get_model_geometry(self):
-        return DocumentGeometry
-
-    def get_model_archive_geometry(self):
-        return ArchiveDocumentGeometry
 
     def get_count_query(self):
         return (
@@ -59,24 +44,6 @@ class MigrateSummits(MigrateDocuments):
             maps_info=document_in.maps_info
         )
 
-    def get_document_archive(self, document_in, version):
-        doc = self.get_document(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
-    def get_document_geometry(self, document_in, version):
-        return dict(
-            document_id=document_in.id,
-            id=document_in.id,
-            version=version,
-            geom=document_in.geom
-        )
-
-    def get_document_geometry_archive(self, document_in, version):
-        doc = self.get_document_geometry(document_in, version)
-        doc['id'] = document_in.document_archive_id
-        return doc
-
     def get_document_locale(self, document_in, version):
         # TODO extract summary
         return dict(
@@ -87,9 +54,6 @@ class MigrateSummits(MigrateDocuments):
             title=document_in.name,
             description=document_in.description
         )
-
-    def get_document_locale_archive(self, document_in, version):
-        return self.get_document_locale(document_in, version)
 
     summit_types = {
         '1': 'summit',    # was 'culmen'
